Day 9

+++ src/day02.rs
@@ -8,8 +8,8 @@ pub fn execute() {
     for i in 1..c {
         for j in 1..c {
             let mut m = read_program(content.clone());
-            m.set_value(1, i as i32);
-            m.set_value(2, j as i32);
+            m.set_value(1, i as i64);
+            m.set_value(2, j as i64);
             m.execute_until_stopped();
             if m.get_value(0) == 19690720 {
                 println!("Noun {}, verb {}, output: {}", i, j, 100 * i + j);
